Refactor CLI feature and add testing
